fix(framework): dts errors that require skipLibCheck
