+++ packages/base/package.json
@@ -5,7 +5,7 @@
   "author": "SAP SE (https://www.sap.com)",
   "license": "Apache-2.0",
   "type": "module",
-  "module": "index.js",
+  "module": "./dist/index.js",
   "keywords": [
     "openui5",
     "sapui5",
@@ -25,8 +25,8 @@
       "default": "./dist/ssr-dom.js"
     },
     ".": {
-      "types": "./src/index.d.ts",
-      "default": "./index.js"
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
     },
     "./dist/*": {
       "production": "./dist/prod/*",
